# openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SnackTrace API",
    "version": "2.0.0",
    "description": "Read-only JSON API for a FoodNoms SQLite database.",
    "license": {
      "name": "ISC",
      "identifier": "ISC"
    }
  },
...
  "paths": {
...
    "/openapi.json": {
      "get": {
        "operationId": "getOpenApi",
        "summary": "Get the OpenAPI document",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "required": [
                    "openapi"
                  ],
                  "properties": {
                    "openapi": {
                      "type": "string",
                      "enum": [
                        "3.1.2"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "path",
                          "message"
                        ],
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "path",
                          "message"
                        ],
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/days/{date}/goals": {
      "get": {
        "operationId": "getDayGoals",
        "summary": "Get a day's goals",
        "tags": [
          "days"
        ],
        "parameters": [
          {
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            },
            "in": "path",
            "name": "date",
            "required": true,
            "description": "Calendar date in strict ISO 8601 YYYY-MM-DD form."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "date",
                    "items"
                  ],
                  "properties": {
                    "date": {
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                      "description": "Calendar date in strict ISO 8601 YYYY-MM-DD form.",
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "goalId",
                          "type",
                          "mode",
                          "lowerBound",
                          "upperBound",
                          "actual",
                          "status",
                          "ratio"
                        ],
                        "properties": {
                          "goalId": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "type": {
                            "type": "string"
                          },
                          "mode": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "maximum"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "minimum"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "range"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "tracking"
                                ]
                              }
                            ]
                          },
                          "lowerBound": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "upperBound": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "actual": {
                            "type": "number"
                          },
                          "status": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "below"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "within"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "above"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "tracking"
                                ]
                              }
                            ]
                          },
                          "ratio": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "path",
                          "message"
                        ],
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "path",
                          "message"
                        ],
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": []
}